sentinel/backend/requirements.txt
fastapi
uvicorn
python-dotenv
supabase
python-telegram-bot
Pillow
pydantic
pydantic-settings
aiohttp
requests
APScheduler
httpx
openai
pytesseract
